Rename stateless package to rest
