+++ rest/handler.go
@@ -10,7 +10,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package stateless
+package rest
 
 import (
 	"context"
